+++ platform/ios/src/MGLMapView.h
@@ -352,6 +352,18 @@ MGL_EXPORT IB_DESIGNABLE
  */
 @property (nonatomic, assign) MGLMapViewPreferredFramesPerSecond preferredFramesPerSecond;
 
+/**
+ A Boolean value indicating whether the map should prefetch tiles.
+ 
+ When this property is set to YES, the map view prefetches loads tiles designed for a
+ low zoom level and displays them until receiving more detailed tiles for the current
+ zoom level. The prefetched tiles typically contain simplified versions of each shape,
+ improving the map view’s perceived performance.
+ 
+ The default value of this property is YES.
+ */
+@property (nonatomic, assign) BOOL prefetchesTiles;
+
 @property (nonatomic) NSArray<NSString *> *styleClasses __attribute__((unavailable("Support for style classes has been removed.")));
 
 - (BOOL)hasStyleClass:(NSString *)styleClass __attribute__((unavailable("Support for style classes has been removed.")));